+++ game/notegraph.cc
@@ -16,7 +16,7 @@ NoteGraph::NoteGraph(VocalTrack const& vocal, NoteGraphScalerPtr const& scaler):
   m_notebargold(findFile("notebargold.svg")), m_notebargold_hl(findFile("notebargold_hi.svg")),
   m_notealpha(0.0f), m_nlTop(0.0, 4.0), m_nlBottom(0.0, 4.0), m_time(), m_scaler(scaler)
 {
-	dimensions.stretch(1.0, 0.5); // Initial dimensions, probably overridden from somewhere
+	dimensions.stretch(1.0f, 0.5f); // Initial dimensions, probably overridden from somewhere
 	m_nlTop.setTarget(m_vocal.noteMax, true);
 	m_nlBottom.setTarget(m_vocal.noteMin, true);
 	for (auto const& n: m_vocal.notes) n.stars.clear(); // Reset stars
@@ -30,18 +30,18 @@ void NoteGraph::reset() {
 }
 
 namespace {
-	void drawNotebar(Texture const& texture, double x, double ybeg, double yend, double w, double h_x, double h_y) {
+	void drawNotebar(Texture const& texture, float x, float ybeg, float yend, float w, float h_x, float h_y) {
 		glutil::VertexArray va;
 		UseTexture tblock(texture);
 
 		// The front cap begins
 		va.texCoord(0.0f, 0.0f).vertex(x, ybeg);
 		va.texCoord(0.0f, 1.0f).vertex(x, ybeg + h_y);
-		if (w >= 2.0 * h_x) {
+		if (w >= 2.0f * h_x) {
 			// Calculate the y coordinates of the middle part
-			double tmp = h_x / w;  // h_x = cap size (because it is a h_x by h_x square)
-			double y1 = (1.0 - tmp) * ybeg + tmp * yend;
-			double y2 = tmp * ybeg + (1.0 - tmp) * yend;
+			float tmp = h_x / w;  // h_x = cap size (because it is a h_x by h_x square)
+			float y1 = (1.0f - tmp) * ybeg + tmp * yend;
+			float y2 = tmp * ybeg + (1.0f - tmp) * yend;
 			// The middle part between caps
 			va.texCoord(0.5f, 0.0f).vertex(x + h_x, y1);
 			va.texCoord(0.5f, 1.0f).vertex(x + h_x, y1 + h_y);
@@ -49,12 +49,12 @@ namespace {
 			va.texCoord(0.5f, 1.0f).vertex(x + w - h_x, y2 + h_y);
 		} else {
 			// Note is too short to even fit caps, crop to fit.
-			double ymid = 0.5 * (ybeg + yend);
+			float ymid = 0.5f * (ybeg + yend);
 			float crop = 0.25f * w / h_x;
-			va.texCoord(crop, 0.0f).vertex(x + 0.5 * w, ymid);
-			va.texCoord(crop, 1.0f).vertex(x + 0.5 * w, ymid + h_y);
-			va.texCoord(1.0f - crop, 0.0f).vertex(x + 0.5 * w, ymid);
-			va.texCoord(1.0f - crop, 1.0f).vertex(x + 0.5 * w, ymid + h_y);
+			va.texCoord(crop, 0.0f).vertex(x + 0.5f * w, ymid);
+			va.texCoord(crop, 1.0f).vertex(x + 0.5f * w, ymid + h_y);
+			va.texCoord(1.0f - crop, 0.0f).vertex(x + 0.5f * w, ymid);
+			va.texCoord(1.0f - crop, 1.0f).vertex(x + 0.5f * w, ymid + h_y);
 		}
 		// The rear cap ends
 		va.texCoord(1.0f, 0.0f).vertex(x + w, yend);
@@ -64,14 +64,14 @@ namespace {
 	}
 }
 
-const double baseLine = -0.2;
-const double pixUnit = 0.2;
+const float baseLine = -0.2f;
+const float pixUnit = 0.2f;
 
 void NoteGraph::draw(double time, Database const& database, Position position) {
 	if (time < m_time) reset();
 	m_time = time;
 	// Update m_songit (which note to start the rendering from)
-	while (m_songit != m_vocal.notes.end() && (m_songit->type == Note::Type::SLEEP || m_songit->end < time - (baseLine + 0.5) / pixUnit)) ++m_songit;
+	while (m_songit != m_vocal.notes.end() && (m_songit->type == Note::Type::SLEEP || m_songit->end < time - (baseLine + 0.5f) / pixUnit)) ++m_songit;
 
 	// Automatically zooming notelines
 	{
@@ -84,25 +84,25 @@ void NoteGraph::draw(double time, Database const& database, Position position) {
 	}
 	switch(position) {
 		case NoteGraph::Position::FULLSCREEN:
-			dimensions.stretch(1.0, 0.50).center();
+			dimensions.stretch(1.0f, 0.50f).center();
 			break;
 		case NoteGraph::Position::TOP:
-			dimensions.stretch(1.0, 0.32).bottom(0.0);
+			dimensions.stretch(1.0f, 0.32f).bottom(0.0f);
 			break;
 		case NoteGraph::Position::BOTTOM:
-			dimensions.stretch(1.0, 0.32).top(0.0);
+			dimensions.stretch(1.0f, 0.32f).top(0.0f);
 			break;
 		case NoteGraph::Position::LEFT:
-			dimensions.stretch(0.50, 0.50).center().left(-0.5);
+			dimensions.stretch(0.50f, 0.50f).center().left(-0.5f);
 			break;
 		case NoteGraph::Position::RIGHT:
-			dimensions.stretch(0.50, 0.50).center().right();
+			dimensions.stretch(0.50f, 0.50f).center().right();
 			break;
 	}
-	m_max = m_nlTop.get() + 7.0;
-	m_min = m_nlBottom.get() - 7.0;
-	m_noteUnit = -dimensions.h() / std::max(48.0 * dimensions.h(), m_max - m_min);
-	m_baseY = -0.5 * (m_min + m_max) * m_noteUnit + dimensions.yc();
+	m_max = m_nlTop.get() + 7.0f;
+	m_min = m_nlBottom.get() - 7.0f;
+	m_noteUnit = -dimensions.h() / std::max(48.0f * dimensions.h(), m_max - m_min);
+	m_baseY = -0.5f * (m_min + m_max) * m_noteUnit + dimensions.yc();
 	m_baseX = baseLine - m_time * pixUnit + dimensions.xc();  // FIXME: Moving in X direction requires additional love (is b0rked now, keep it centered at zero)
 
 	// Fading notelines handing
@@ -116,35 +116,35 @@ void NoteGraph::draw(double time, Database const& database, Position position) {
 	if (config["game/pitch"].b()) drawWaves(database);
 
 	// Draw a star for well sung notes
-	for (auto it = m_songit; it != m_vocal.notes.end() && it->begin < m_time - (baseLine - 0.5) / pixUnit; ++it) {
+	for (auto it = m_songit; it != m_vocal.notes.end() && it->begin < m_time - (baseLine - 0.5f) / pixUnit; ++it) {
 		float player_star_offset = 0;
 		for (std::vector<Color>::const_iterator it_col = it->stars.begin(); it_col != it->stars.end(); ++it_col) {
-			double x = m_baseX + it->begin * pixUnit + m_noteUnit; // left x coordinate: begin minus border (side borders -noteUnit wide)
-			double w = (it->end - it->begin) * pixUnit - m_noteUnit * 2.0; // width: including borders on both sides
+			float x = m_baseX + it->begin * pixUnit + m_noteUnit; // left x coordinate: begin minus border (side borders -noteUnit wide)
+			float w = (it->end - it->begin) * pixUnit - m_noteUnit * 2.0f; // width: including borders on both sides
 			float hh = -m_noteUnit;
-			float centery = m_baseY + (it->note + 0.4) * m_noteUnit; // Star is 0.4 notes higher than current note
-			float centerx = x + w - (player_star_offset + 1.2) * hh; // Star is 1.2 units from end
+			float centery = m_baseY + (it->note + 0.4f) * m_noteUnit; // Star is 0.4 notes higher than current note
+			float centerx = x + w - (player_star_offset + 1.2f) * hh; // Star is 1.2 units from end
 			float rot = std::remainder(time * 5.0, TAU); // They rotate!
 			bool smallerNoteGraph = ((position == NoteGraph::Position::TOP) || (position == NoteGraph::Position::BOTTOM));
-			float zoom = (std::abs((rot-180) / 360.0f) * 0.8f + 0.6f) * (smallerNoteGraph ? 2.3 : 2.0) * hh;
+			float zoom = (std::abs((rot-180) / 360.0f) * 0.8f + 0.6f) * (smallerNoteGraph ? 2.3f : 2.0f) * hh;
 			using namespace glmath;
 			Transform trans(translate(vec3(centerx, centery, 0.0f)) * rotate(rot, vec3(0.0f, 0.0f, 1.0f)));
 			{
 				ColorTrans c(Color(it_col->r, it_col->g, it_col->b, it_col->a));
-				m_star_hl.draw(Dimensions().stretch(zoom*1.2, zoom*1.2).center().middle(), TexCoords());
+				m_star_hl.draw(Dimensions().stretch(zoom*1.2f, zoom*1.2f).center().middle(), TexCoords());
 			}
 			m_star.draw(Dimensions().stretch(zoom, zoom).center().middle(), TexCoords());
-			player_star_offset += 0.8;
+			player_star_offset += 0.8f;
 		}
 	}
 }
 
 void NoteGraph::drawNotes() {
 	// Draw note lines
-	m_notelines.draw(Dimensions().stretch(dimensions.w(), (m_max - m_min - 13) * m_noteUnit).middle(dimensions.xc()).center(dimensions.yc()), TexCoords(0.0, (-m_min - 7.0) / 12.0f, 1.0, (-m_max + 6.0) / 12.0f));
+	m_notelines.draw(Dimensions().stretch(dimensions.w(), (m_max - m_min - 13) * m_noteUnit).middle(dimensions.xc()).center(dimensions.yc()), TexCoords(0.0f, (-m_min - 7.0f) / 12.0f, 1.0f, (-m_max + 6.0f) / 12.0f));
 
 	// Draw notes
-	for (auto it = m_songit; it != m_vocal.notes.end() && it->begin < m_time - (baseLine - 0.5) / pixUnit; ++it) {
+	for (auto it = m_songit; it != m_vocal.notes.end() && it->begin < m_time - (baseLine - 0.5f) / pixUnit; ++it) {
 		if (it->type == Note::Type::SLEEP) continue;
 		double alpha = it->power;
 		Texture* t1;
@@ -159,26 +159,26 @@ void NoteGraph::drawNotes() {
 			case Note::Type::RAP: //handle RAP notes like freestyle for now
 			{
 				Dimensions dim;
-				dim.middle(m_baseX + 0.5 * (it->begin + it->end) * pixUnit).center(m_baseY + it->note * m_noteUnit).stretch((it->end - it->begin) * pixUnit, -m_noteUnit * 12.0);
+				dim.middle(m_baseX + 0.5f * (it->begin + it->end) * pixUnit).center(m_baseY + it->note * m_noteUnit).stretch((it->end - it->begin) * pixUnit, -m_noteUnit * 12.0f);
 				float xoffset = 0.1 * m_time / m_notebarfs.dimensions.ar();
-				m_notebarfs.draw(dim, TexCoords(xoffset, 0.0, xoffset + dim.ar() / m_notebarfs.dimensions.ar(), 1.0));
-				if (alpha > 0.0) {
+				m_notebarfs.draw(dim, TexCoords(xoffset, 0.0f, xoffset + dim.ar() / m_notebarfs.dimensions.ar(), 1.0f));
+				if (alpha > 0.0f) {
 					float xoffset = rand() / double(RAND_MAX);
-					m_notebarfs_hl.draw(dim, TexCoords(xoffset, 0.0, xoffset + dim.ar() / m_notebarfs_hl.dimensions.ar(), 1.0));
+					m_notebarfs_hl.draw(dim, TexCoords(xoffset, 0.0f, xoffset + dim.ar() / m_notebarfs_hl.dimensions.ar(), 1.0f));
 				}
 			}
 			continue;
 		  default: throw std::logic_error("Unknown note type: don't know how to render");
 		}
-		double x = m_baseX + it->begin * pixUnit + m_noteUnit; // left x coordinate: begin minus border (side borders -noteUnit wide)
+		float x = m_baseX + it->begin * pixUnit + m_noteUnit; // left x coordinate: begin minus border (side borders -noteUnit wide)
 		double bar_height = barHeight();
 		double ybeg = m_baseY + (it->notePrev +bar_height) * m_noteUnit; // top y coordinate (on the one higher note line)
 		double yend = m_baseY + (it->note +bar_height) * m_noteUnit; // top y coordinate (on the one higher note line)
-		double w = (it->end - it->begin) * pixUnit - m_noteUnit * 2.0; // width: including borders on both sides
-		double h_x = -m_noteUnit * 2.0; // height: 0.5 border + 1.0 bar + 0.5 border = 2.0
-		double h_y = h_x * bar_height; //
+		float w = (it->end - it->begin) * pixUnit - m_noteUnit * 2.0f; // width: including borders on both sides
+		float h_x = -m_noteUnit * 2.0f; // height: 0.5 border + 1.0 bar + 0.5 border = 2.0
+		float h_y = h_x * bar_height; //
 		drawNotebar(*t1, x, ybeg, yend, w, h_x, h_y);
-		if (alpha > 0.0) {
+		if (alpha > 0.0f) {
 			ColorTrans c(Color::alpha(alpha));
 			drawNotebar(*t2, x, ybeg, yend, w, h_x, h_y);
 		}
@@ -221,14 +221,14 @@ void NoteGraph::drawWaves(Database const& database) {
 		double oldval = getNaN();
 		glutil::VertexArray va;
 		auto noteIt = m_vocal.notes.begin();
-		glmath::vec4 c(player.m_color.r, player.m_color.g, player.m_color.b, 1.0);
+		glmath::vec4 c(player.m_color.r, player.m_color.g, player.m_color.b, 1.0f);
 		for (; idx < endIdx; ++idx, t += Engine::TIMESTEP) {
 			double const freq = pitch[idx].first;
 			// If freq is NaN, we have nothing to process
 			if (freq != freq) { oldval = getNaN(); tex = texOffset; continue; }
 			tex += freq * 0.001; // Wave phase (texture coordinate)
 			if (idx < beginIdx) continue; // Skip graphics rendering if out of screen
-			double x = -0.2 + (t - m_time) * pixUnit;
+			float x = -0.2f + (t - m_time) * pixUnit;
 			// Find the currently active note(s)
 			while (noteIt != m_vocal.notes.end() && (noteIt->type == Note::Type::SLEEP || t > noteIt->end)) ++noteIt;
 			auto notePrev = noteIt;
@@ -242,7 +242,7 @@ void NoteGraph::drawWaves(Database const& database) {
 			// Now val contains the active note value. The following calculates note value for current freq:
 			val += Note::diff(val, MusicalScale(m_vocal.scale).setFreq(freq).getNote());
 			// Graphics positioning & animation:
-			double y = m_baseY + val * m_noteUnit;
+			float y = m_baseY + val * m_noteUnit;
 			double thickness = clamp(1.0 + pitch[idx].second / 60.0) + 0.5;
 			thickness *= NoteGraph::waveThickness() * (1.0 + 0.2 * std::sin(tex - 2.0 * texOffset)); // Further animation :)
 			thickness *= -m_noteUnit;
@@ -251,8 +251,8 @@ void NoteGraph::drawWaves(Database const& database) {
 			// Add a point or a pair of points
 			if (!va.size()) va.texCoord(tex, 0.5f).color(c).vertex(x, y);
 			else {
-				va.texCoord(tex, 0.0f).color(c).vertex(x, y - thickness);
-				va.texCoord(tex, 1.0f).color(c).vertex(x, y + thickness);
+				va.texCoord(tex, 0.0f).color(c).vertex(x, y - static_cast<float>(thickness));
+				va.texCoord(tex, 1.0f).color(c).vertex(x, y + static_cast<float>(thickness));
 			}
 			oldval = val;
 		}
